+++ packages/plugin-packages/mashroom-session-provider-redis/src/mashroom-bootstrap.js
@@ -9,6 +9,9 @@ const bootstrap: MashroomSessionStoreProviderPluginBootstrapFunction = async (pl
     const logger = pluginContextHolder.getPluginContext().loggerFactory('mashroom.session.provider.redis');
     logger.info('Using Redis client options:', pluginConfig);
     const client = createClient(pluginConfig);
+    client.on('error', (err: any) => {
+        logger.error('Redis store error:', err);
+    });
     const options = {
         client,
     };